Use ddtrace for span creation (#48)
Before, we created and managed traces in troncos using the OTEL libraries. The downside of this was that traces were not exported nicely to DD and third-party libraries were poorly traced.

This PR flips this around, we use ddtrace as our main method of tracing and support shipping those traces to OTEL collectors. That way, we keep all the information that people are used to with DD, while also supporting OTEL.