✨ feature(structlog): Add extra_processors and extra_loggers
Adds two optional kwargs to `configure_structlog`,

`extra_processors` takes a list of extra processors which will be injected just before `format_exc_info`.

`extra_loggers` takes a dict which will be unpacked as part of the `loggers` directive in the core logging config.

This enables apps to add their own logging processors in addition to those defined in `troncos`, and to customize log level and propagation of logger paths.